requirements.txt
Flask==2.1.1
Pillow>=10.0.0
tqdm==4.62.3
